Fixed ints used for karma levels to int64

+++ internal/karma/karma.go
@@ -2,9 +2,9 @@ package karma
 
 import "github.com/connorkuehl/popple/internal/parse"
 
-func Parse(text string) map[string]int {
+func Parse(text string) map[string]int64 {
 	subjects := parse.Subjects(text)
-	levels := make(map[string]int)
+	levels := make(map[string]int64)
 
 	for _, s := range subjects {
 		levels[s.Name] += s.Karma
